chore(refactor): move OGC SchemaCollection to ogc-api-feature/types.ts (refs #47)
